pyproject.toml
[project]
name = "yt-streamlit"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-generativeai>=0.8.3",
    "pocketbase>=0.14.0",
    "python-dotenv>=1.0.1",
    "streamlit>=1.41.1",
    "youtube-transcript-api>=0.6.3",
]
